Fix equality implementation for Java.
Implement abs/1 in Mercury for the C# backend so that its behaviour is
consistent with the other backends.

src/int32.m:
src/int64.m:
src/uint32.m:
src/uint64.m:
    As above.

src/uint64.m:
    Add an explanation for the cast in the implementation of to_binary_string/1.

tests/arith_int32.{m,exp}:
tests/arith_int64.{m,exp}:
    Test unary operators (and the abs function).

tests/run_tests.sh:
    Fix the order in which expected outputs and results are diffed.

+++ src/uint32.m
@@ -193,7 +193,7 @@
     uint32_equal(A::in, B::in),
     [will_not_call_mercury, promise_pure, thread_safe],
 "
-    SUCCESS_INDICATOR = (A == B);
+    SUCCESS_INDICATOR = (A.intValue() == B.intValue());
 ").
 
 uint32_compare(Result, A, B) :-
